Bug 1201505 - Pre: Extracted InnerStrokedView into it's own file
